Fix Db operation bug. Add check null and close the cursor

+++ src/sg/gov/nhb/ihuayu/activity/operation/IhuayuOperationImpl.java
@@ -1,5 +1,5 @@
 /**
- * 
+ *
  */
 
 package sg.gov.nhb.ihuayu.activity.operation;
@@ -20,9 +20,12 @@ import sg.gov.nhb.ihuayu.activity.db.entity.QueryType;
 import sg.gov.nhb.ihuayu.activity.db.entity.Scenario;
 import sg.gov.nhb.ihuayu.activity.db.entity.ScenarioDialog;
 import sg.gov.nhb.ihuayu.activity.rest.AudioPlayer;
+
+import android.app.Dialog;
 import android.content.ContentValues;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
+import android.text.TextUtils;
 import android.util.Log;
 
 /**
@@ -31,7 +34,7 @@ import android.util.Log;
 public class IhuayuOperationImpl {
     private static final String TAG = "iHuayu:IhuayuOperationImpl";
     private SQLiteDatabase db = null;
-    DBSqlite manager;
+    private DBSqlite manager;
 
     public IhuayuOperationImpl(DBSqlite manager) {
         this.db = manager.getSqlDB();
@@ -49,26 +52,39 @@ public class IhuayuOperationImpl {
     }
 
     public List<Scenario> queryScenario(String condition, String[] params) {
-        Cursor result = this.db.rawQuery(condition, params);
-        return OperationUtils.cursorToScenario(result);
+        Cursor cursor = null;
+        List<Scenario> list = null;
+        try {
+            if (db != null) {
+                cursor = this.db.rawQuery(condition, params);
+                list = OperationUtils.cursorToScenario(cursor);
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+        } finally {
+            cursor.close();
+        }
+        return list;
     }
 
     public void insertDictionary(List<ContentValues> params) {
         // db.execSQL("CREATE TABLE name (id INTEGER, language_dir VARCHAR(10), keyword VARCHAR(1024),keyword_length INTEGER, src VARCHAR(1024), destination VARCHAR(1024), chinese_audio_link VARCHAR(1024), chinese_py_with_tone TEXT, dict_category CHAR(50), sample_sentence_EN VARCHAR(1024), sample_sentence_CN VARCHAR(1024), sample_sentence_PY VARCHAR(1024), sample_sentence_CN_Audio_link VARCHAR(1024))");
-        for (ContentValues values : params) {
-            // long result = this.db.insert("dictionary", null, values);
-            this.db.insert("dictionary", null, values);
+        if (db != null) {
+            for (ContentValues values : params) {
+                // long result = this.db.insert("dictionary", null, values);
+                this.db.insert("dictionary", null, values);
+            }
         }
     }
 
     public long insertDictionary(ContentValues param) {
-        if (param != null) {
+        if (param != null && db != null) {
             String id = param.getAsString("id");
             if ("true".equalsIgnoreCase(param.getAsString("deleted"))) {
                 deleteDictionary(param.getAsString("id"));
             } else {
-                int updateCount = this.db.update("dictionary", param, "id=?", new String[] {
-                    id
+                int updateCount = this.db.update("dictionary", param, "id=?", new String[]{
+                        id
                 });
                 if (updateCount == 0) {
                     this.db.insert("dictionary", null, param);
@@ -85,12 +101,14 @@ public class IhuayuOperationImpl {
     }
 
     public void deleteDictionary(String id) {
-        this.db.delete("Favorites", "Dictionary_ID=?", new String[] {
-            id
-        });
-        this.db.delete("dictionary", "id=?", new String[] {
-            id
-        });
+        if (db != null) {
+            this.db.delete("Favorites", "Dictionary_ID=?", new String[]{
+                    id
+            });
+            this.db.delete("dictionary", "id=?", new String[]{
+                    id
+            });
+        }
     }
 
     public void insertIntoScenarioDialogKeyWord(
@@ -124,40 +142,46 @@ public class IhuayuOperationImpl {
     }
 
     public void deleteScenario(String id) {
-        // Get Dialog
-        List<ScenarioDialog> dialogList = queryDialog(
-                "select * from Dialog where title_id = ? ", new String[] {
+        if (db != null) {
+            // Get Dialog
+            List<ScenarioDialog> dialogList = queryDialog(
+                    "select * from Dialog where title_id = ? ", new String[]{
                     id
+            });
+            for (ScenarioDialog dialog : dialogList) {
+                db.delete("Dialog_Keyword", "Dialog_ID=?", new String[]{
+                        dialog.getId()
                 });
-        for (ScenarioDialog dialog : dialogList) {
-            db.delete("Dialog_Keyword", "Dialog_ID=?", new String[] {
-                dialog.getId()
+            }
+            db.delete("Dialog", "Title_ID=?", new String[]{
+                    id
+            });
+            db.delete("Scenario_Category", "Title_ID=?", new String[]{
+                    id
             });
         }
-        db.delete("Dialog", "Title_ID=?", new String[] {
-            id
-        });
-        db.delete("Scenario_Category", "Title_ID=?", new String[] {
-            id
-        });
     }
 
     public void deleteDialog(String id) {
-        // Get Dialog
-        db.delete("Dialog", "Dialog_ID=?", new String[] {
-            id
-        });
+        if (db != null) {
+            // Get Dialog
+            db.delete("Dialog", "Dialog_ID=?", new String[]{
+                    id
+            });
+        }
     }
 
     public void deleteDialogKeyword(String id) {
-        // Get Dialog
-        db.delete("Dialog_Keyword", "ID=?", new String[] {
-            id
-        });
+        if (db != null) {
+            // Get Dialog
+            db.delete("Dialog_Keyword", "ID=?", new String[]{
+                    id
+            });
+        }
     }
 
     public void insertIntoScenarioDialogKeyWord(ContentValues contentValues,
-            HashMap<ContentValues, List<ContentValues>> dialogKeywordsMap) {
+                                                HashMap<ContentValues, List<ContentValues>> dialogKeywordsMap) {
         if (contentValues != null && contentValues.size() > 0) {
             // insertTable("Scenario_Category", contentValues);
             if ("true".equalsIgnoreCase(contentValues.getAsString("deleted"))) {
@@ -197,12 +221,12 @@ public class IhuayuOperationImpl {
     }
 
     public int updateAndInsertScenario(ContentValues contentValues, String tableName,
-            String primeryKeyId) {
-        if (contentValues != null) {
+                                       String primeryKeyId) {
+        if (contentValues != null && db != null) {
             String id = contentValues.getAsString(primeryKeyId);
             int updateCount = this.db.update(tableName, contentValues, primeryKeyId + "=?",
-                    new String[] {
-                        id
+                    new String[]{
+                            id
                     });
             if (updateCount == 0) {
                 this.db.insert(tableName, null, contentValues);
@@ -217,27 +241,40 @@ public class IhuayuOperationImpl {
     }
 
     public long insertToBookmark(String dictionaryID) {
-        try {
-            ContentValues contentValues = new ContentValues();
-            contentValues.put("Dictionary_ID", dictionaryID);
-            contentValues.put("Update_date", getStandardTimeStr());
-            return this.db.insert("Favorites", null, contentValues);
-        } catch (Exception e) {
+        if (db != null) {
+            try {
+                ContentValues contentValues = new ContentValues();
+                contentValues.put("Dictionary_ID", dictionaryID);
+                contentValues.put("Update_date", getStandardTimeStr());
+                return this.db.insert("Favorites", null, contentValues);
+            } catch (Exception e) {
+                return -1;
+            }
+        } else {
             return -1;
         }
     }
 
     public List<Dictionary> getAllBookmarks() {
-        try {
-            // TODO. maybe have performance issue.
-            Cursor result = this.db
-                    .rawQuery(
-                            "select rowid, * from dictionary where id in (select Dictionary_ID from Favorites)",
-                            null);
-            return OperationUtils.cursorToDictionary(result);
-        } catch (Exception e) {
-            return null;
+        List<Dictionary> list = null;
+        Cursor cursor = null;
+        if (db != null) {
+            try {
+                // TODO. maybe have performance issue.
+                cursor = this.db
+                        .rawQuery(
+                                "select rowid, * from dictionary where id in (select Dictionary_ID from Favorites)",
+                                null);
+                list = OperationUtils.cursorToDictionary(cursor);
+            } catch (Exception e) {
+                e.printStackTrace();
+            } finally {
+                if (cursor != null) {
+                    cursor.close();
+                }
+            }
         }
+        return list;
     }
 
     public int getBookmark(int dictionaryId) {
@@ -245,8 +282,8 @@ public class IhuayuOperationImpl {
             Log.d(TAG, "[getBookmark] + Begin");
             // TODO. maybe have performance issue.
             Cursor result = this.db.rawQuery("select * from Favorites where Dictionary_ID = ?",
-                    new String[] {
-                        dictionaryId + ""
+                    new String[]{
+                            dictionaryId + ""
                     });
             // return result.getInt(result.getColumnIndex("Dictionary_ID"));
             Log.d(TAG, "[getBookmark] Cursor Count = " + result.getCount());
@@ -257,9 +294,12 @@ public class IhuayuOperationImpl {
     }
 
     public int removeFromFavorites(int dictionaryId) {
-        int result = this.db.delete("Favorites", "Dictionary_ID = ?", new String[] {
-                dictionaryId + ""
-        });
+        int result = 0;
+        if (db != null) {
+            result = this.db.delete("Favorites", "Dictionary_ID = ?", new String[]{
+                    dictionaryId + ""
+            });
+        }
         return result;
     }
 
@@ -277,34 +317,89 @@ public class IhuayuOperationImpl {
     }
 
     public long insertTable(String name, ContentValues contentValues) {
-        long result = this.db.insert(name, null, contentValues);
+        long result = 0;
+        if (db != null) {
+            result = this.db.insert(name, null, contentValues);
+        }
         return result;
     }
 
     public List<ScenarioDialog> queryDialog(String condition, String[] params) {
-        Cursor result = this.db.rawQuery(condition, params);
-        return OperationUtils.cursorToDialog(result);
+        Cursor cursor = null;
+        List<ScenarioDialog> list = null;
+        if (db != null) {
+            try {
+                cursor = this.db.rawQuery(condition, params);
+                list = OperationUtils.cursorToDialog(cursor);
+            } catch (Exception e) {
+                e.printStackTrace();
+            } finally {
+                if (cursor != null) {
+                    cursor.close();
+                }
+            }
+        }
+        return list;
     }
 
     public List<Dictionary> queryDictionary(String condition, String[] params) {
-        Cursor result = this.db.rawQuery(condition, params);
-        return OperationUtils.cursorToDictionary(result);
+        Cursor cursor = null;
+        List<Dictionary> list = null;
+        if (db != null) {
+            try {
+                cursor = this.db.rawQuery(condition, params);
+                list = OperationUtils.cursorToDictionary(cursor);
+            } catch (Exception e) {
+                e.printStackTrace();
+            } finally {
+                if (cursor != null) {
+                    cursor.close();
+                }
+            }
+        }
+        return list;
     }
 
     public List<DialogKeywords> queryDialogKeywords(String condition, String[] params) {
-        Cursor result = this.db.rawQuery(condition, params);
-        return OperationUtils.cursorToDialogKeywords(result);
+        Cursor cursor = null;
+        List<DialogKeywords> list = null;
+        if (db != null) {
+            try {
+                cursor = this.db.rawQuery(condition, params);
+                list = OperationUtils.cursorToDialogKeywords(cursor);
+            } catch (Exception e) {
+                e.printStackTrace();
+            } finally {
+                if (cursor != null) {
+                    cursor.close();
+                }
+            }
+        }
+        return list;
     }
 
     public List<Dictionary> searchDictionary(String language_dir, String keyword) {
-        db.execSQL("create index IF NOT EXISTS dictinaryIndex ON dictionary (language_dir, keyword)");
-        Cursor result = db
-                .rawQuery(
+        List<Dictionary> list = null;
+        if (db != null) {
+            db.execSQL("create index IF NOT EXISTS dictinaryIndex ON dictionary (language_dir, keyword)");
+            Cursor cursor = null;
+            try {
+                cursor = db.rawQuery(
                         "select rowid, * from dictionary where language_dir = ? and keyword like ? order by rowid limit 20",
-                        new String[] {
+                        new String[]{
                                 language_dir, keyword + "%"
                         });
-        List<Dictionary> list = OperationUtils.cursorToDictionary(result);
+                list = OperationUtils.cursorToDictionary(cursor);
+            } catch (Exception e) {
+                e.printStackTrace();
+            } finally {
+                if (cursor != null) {
+                    cursor.close();
+                }
+            }
+
+
+        }
         // Collections.sort(list);
         return list;
     }
@@ -312,9 +407,9 @@ public class IhuayuOperationImpl {
     public FuzzyResult fuzzySearch(String language_dir, String keyword) {
         FuzzyResult result = new FuzzyResult();
         List<Dictionary> exactResult = queryDictionary(
-                "select rowid, * from dictionary where keyword like ? ", new String[] {
-                    keyword
-                });
+                "select rowid, * from dictionary where keyword like ? ", new String[]{
+                keyword
+        });
         if (exactResult != null && exactResult.size() > 0) {
             result.setDictionaryList(exactResult);
             result.setExactResult(true);
@@ -326,36 +421,38 @@ public class IhuayuOperationImpl {
 
         // Select from Scenario
         List<Scenario> scenarioList = null;
-        if(QueryType.EN.getName().equals(language_dir)) {
-        	scenarioList = queryScenario(
-                "SELECT * FROM Scenario_Category WHERE Title_EN LIKE ? ", new String[] {
+        if (QueryType.EN.getName().equals(language_dir)) {
+            scenarioList = queryScenario(
+                    "SELECT * FROM Scenario_Category WHERE Title_EN LIKE ? ", new String[]{
                     "%" + keyword + "%"
-                });
-        }else {
-        	scenarioList = queryScenario(
-                    "SELECT * FROM Scenario_Category WHERE Title_CN LIKE ? ", new String[] {
-                        "%" + keyword + "%"
-                    });
+            });
+        } else {
+            scenarioList = queryScenario(
+                    "SELECT * FROM Scenario_Category WHERE Title_CN LIKE ? ", new String[]{
+                    "%" + keyword + "%"
+            });
         }
         result.setScenarioList(scenarioList);
         return result;
     }
 
     public String getLastUpdateTime() {
-        String dataTime = "";
-        Cursor result = this.db.rawQuery("select Last_Update from Information where version = ?",
-                new String[] {
-                    OperationUtils.getCurrentApplicationVersion(this.manager.getContext())
-                });
-        if (result != null) {
-            if (result.moveToNext()) {
-                dataTime = result.getString(result.getColumnIndex("Last_Update"));
+        String dataTime = null;
+        if (db != null) {
+            Cursor result = this.db.rawQuery("select Last_Update from Information where version = ?",
+                    new String[]{
+                            OperationUtils.getCurrentApplicationVersion(this.manager.getContext())
+                    });
+            if (result != null) {
+                if (result.moveToNext()) {
+                    dataTime = result.getString(result.getColumnIndex("Last_Update"));
+                }
+            } else {
+                Log.w(TAG, "Cannot query last update time");
             }
-        } else {
-            Log.w(TAG, "Cannot query last update time");
         }
 
-        if (dataTime == null || dataTime.length() <= 0) {
+        if (TextUtils.isEmpty(dataTime)) {
             dataTime = "2011-04-01T01S01S01";
         }
         return dataTime;
@@ -367,36 +464,13 @@ public class IhuayuOperationImpl {
             contentValues.put("Version",
                     OperationUtils.getCurrentApplicationVersion(this.manager.getContext()));
             contentValues.put("Last_Update", getUpdatesTimeStr());
-            return this.db.insert("Information", null, contentValues);
+            if (db != null) {
+                return this.db.insert("Information", null, contentValues);
+            } else {
+                return -1;
+            }
         } catch (Exception e) {
             return -1;
         }
     }
-
-    // public String getCancelTime() {
-    // String dataTime = "";
-    // Cursor result = this.db.rawQuery(
-    // "select Last_Update from Information where version = 'cancel'", null);
-    // if (result != null) {
-    // if (result.moveToNext()) {
-    // dataTime = result.getString(result.getColumnIndex("Last_Update"));
-    // }
-    // }
-    // if (dataTime == null || dataTime.length() <= 0) {
-    // dataTime = null;
-    // }
-    // return dataTime;
-    // }
-
-    // public long updateCancelTime() {
-    // try {
-    // ContentValues contentValues = new ContentValues();
-    // contentValues.put("Version", "cancel");
-    // contentValues.put("Last_Update", getStandardTimeStr());
-    // return this.db.insert("Information", null, contentValues);
-    // } catch (Exception e) {
-    // return -1;
-    // }
-    // }
-
 }
